Release records after they've been consumed in streaming mode (#127)
Fixes #126

When `Client#query` is called with the `streaming: true` option, records
do not currently get released to GC as they are consumed. This change
replaces a consumed partition with a single symbol to allow consumed
partitions to be freed.

This technically represents a minor breaking change, as previously the
streaming `Result` object could be iterated multiple times. I think that
is unexpected for a streaming response anyway, so it's unlikely to be
problematic.