Actually send the query timeout to snowflake, allow setting per query (#136)
This addresses #128 - we'll now pass the timeout parameter in the query
to snowflake.

Additionally, seemed to me we should probably allow this per-query as
well so you can have a default and allow some long/slow query to have a
different value. Had to pass the setting around as a method param around
to maintain thread safety.